Fix application crashing bug
Program crashes when invalid date is edited in .json file for Duty.

The program couldn't handle the error gracefully and the application
couldn't start.

JsonAdaptedPerson.java toModelType() function is modified to handle
the issue gracefully. The function checks for the validity of the
dates and throws exception if the dates are invalid.